pdf_ai_assistant/requirements.txt
fastapi==0.99.1
starlette==0.27.0
unicorn==2.0.1.post1
PyPDF2==3.0.1
python-dotenv==1.0.0
langchain==0.0.260